+++ src/pipecatcloud/cli/commands/agent.py
@@ -157,13 +157,6 @@ async def status(
             str(data.get("updatedAt", "N/A")),
         )
 
-        deployment_panel = Panel(
-            deployment_table,
-            title="[bold]Deployment info:[/bold]",
-            title_align="left",
-            border_style="dim",
-        )
-
         # Autoscaling info
         autoscaling_data = data.get("autoScaling", None)
         if autoscaling_data:
@@ -196,14 +189,14 @@ async def status(
         console.print(
             Panel(
                 Group(
+                    deployment_table,
+                    autoscaling_panel if autoscaling_panel else "",
+                    table,
                     Panel(
                         f"[{color}]Health: {'Ready' if data['ready'] else 'Stopped'}[/]",
                         border_style="green" if data['ready'] else "yellow",
                         expand=False,
-                    ),
-                    table,
-                    deployment_panel,
-                    autoscaling_panel if autoscaling_panel else ""
+                    )
                 ),
                 title=f"Status for agent [bold]{agent_name}[/bold]",
                 title_align="left",
@@ -226,7 +219,7 @@ async def sessions(
 ):
     org = organization or config.get("org")
 
-    with Live(console.status(f"[dim]Looking up agent with name {agent_name}[/dim]", spinner="dots")) as live:
+    with Live(console.status(f"[dim]Looking up agent with name '{agent_name}'[/dim]", spinner="dots")) as live:
         data, error = await API.agent(agent_name=agent_name, org=org, live=live)
 
         live.stop()
@@ -237,6 +230,9 @@ async def sessions(
         if not data:
             console.error(f"No deployment data found for agent with name '{agent_name}'")
             return typer.Exit()
+
+        console.print(
+            "[yellow]Please note: this method is currently work in progress and will be updated in the future with more information[/yellow]")
 
         if data.get('activeSessionCount', 0) > 0:
             console.success(
@@ -294,10 +290,21 @@ async def delete(
         "-o",
         help="Organization to delete agent from",
     ),
+    force: bool = typer.Option(
+        False,
+        "--force",
+        "-f",
+        help="Bypass prompt for confirmation",
+    ),
 ):
     org = organization or config.get("org")
 
-    with console.status(f"Deleting agent: [bold]'{agent_name}'[/bold]", spinner="dots"):
+    if not force:
+        if not await questionary.confirm("Are you sure you want to delete this agent? Note: active sessions will not be interrupted and will continue to run until completion.").ask_async():
+            console.print("[bold]Aborting delete request[/bold]")
+            return typer.Exit(1)
+
+    with console.status(f"[dim]Deleting agent: [bold]'{agent_name}'[/bold][/dim]", spinner="dots"):
         data, error = await API.agent_delete(agent_name=agent_name, org=org)
 
         if error:
@@ -328,7 +335,7 @@ async def deployments(
     error_code = None
 
     try:
-        with console.status(f"Fetching deployments for agent: [bold]'{agent_name}'[/bold]", spinner="dots"):
+        with console.status(f"[dim]Fetching deployments for agent: [bold]'{agent_name}'[/bold][/dim]", spinner="dots"):
             async with aiohttp.ClientSession() as session:
                 response = await session.get(
                     f"{API.construct_api_url('services_deployments_path').format(org=org, service=agent_name)}",
@@ -337,7 +344,9 @@ async def deployments(
             if response.status != 200:
                 error_code = str(response.status)
                 response.raise_for_status()
+
             data = await response.json()
+            print(data)
 
             table = Table(
                 show_header=True,
@@ -404,7 +413,15 @@ async def start(
         help="Create a Daily WebRTC session for the agent",
         rich_help_panel="Start Configuration",
     ),
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization which the agent belongs to",
+    ),
 ):
+    org = organization or config.get("org")
+
     default_public_api_key = api_key or config.get("default_public_key")
     default_public_api_key_name = "CLI provided" if api_key else config.get(
         "default_public_key_name")
@@ -432,11 +449,24 @@ async def start(
             title_align="left",
             border_style="yellow",
         ))
-        if not await questionary.confirm("Are you sure you want to start this agent?").ask_async():
+        if not await questionary.confirm("Are you sure you want to start an active session for this agent?").ask_async():
             console.print("[bold]Aborting start request[/bold]")
             return typer.Exit(1)
 
-    with Live(console.status(f"Sending start request with key: {default_public_api_key_name}", spinner="dots"), refresh_per_second=4) as live:
+    with Live(console.status(f"[dim]Checking agent health...[/dim]", spinner="dots"), refresh_per_second=4) as live:
+
+        health_data, error = await API.agent(agent_name=agent_name, org=org, live=live)
+        if not health_data or not health_data['ready']:
+            live.stop()
+            console.error(
+                f"Agent '{agent_name}' does not exist or is not in a health state. Please check the agent status with [bold cyan]{PIPECAT_CLI_NAME} agent status {agent_name}[/bold cyan]")
+            return typer.Exit(1)
+
+        live.update(
+            console.status(
+                f"[dim]Agent '{agent_name}' is healthy, sending start request...[/dim]",
+                spinner="dots"))
+
         data, error = await API.start_agent(agent_name=agent_name, api_key=default_public_api_key, use_daily=use_daily, data=data, live=live)
 
         if error:
@@ -449,7 +479,6 @@ async def start(
                 subtitle=f"[white dim]Deploy an agent with[/white dim] [bold cyan]{PIPECAT_CLI_NAME} deploy[/bold cyan]")
             return typer.Exit(1)
 
-        live.update(console.status(f"Agent '{agent_name}' started successfully", spinner="dots"))
         live.stop()
 
         message = f"Agent '{agent_name}' started successfully"