+++ src/pipecatcloud/cli/entry_point.py
@@ -61,7 +61,7 @@ entrypoint_cli_typer = typer.Typer(
 
 
 @entrypoint_cli_typer.callback()
-def pipecat(
+def cli(
     ctx: typer.Context,
     _version: bool = typer.Option(None, "--version", callback=version_callback, help="CLI version"),
     _config: bool = typer.Option(None, "--config", callback=config_callback, help="CLI config"),